Add placeholder rows to Case Insights table so that all rollup levels are covered (Recidiviz/recidiviz-data#36671)
## Description of the change

As described in Recidiviz/recidiviz-data#36670, not all possible roll-up values are represented
in the case insights table, which is a problem when category overrides
are used -- the override logic searches for a row in the table matching
the newly overridden combination of categories (e.g. Violent, Drug
offense) to find the corresponding recidivism rates, but if no such row
exists then it's unable to retrieve anything.

This PR adds [PLACEHOLDER] cases with all combinations of the
violent/drug/sex offense attributes to the underlying data before
performing rollups, guaranteeing that any combination will have a
corresponding rollup in the table.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#36670

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: ca9ff7b3be1674c6daa07baab6cafadb85df1428