[Insights] restrict insights_caseload_category_sessions to insights states (Recidiviz/recidiviz-data#38757)
## Description of the change

See slack thread
(https://recidiviz.slack.com/archives/C028X32LRH7/p1740502073285569) for
context. This restricts the view only to insights states since it's only
used specifically to funnel into the product.

Loaded just this view into a sandbox at
`dana_caseload_category_sessions`. It outputted 11,641,007 rows compared
with 12,496,531 in the prod version (the staging version didn't get
materialized), so this fix might not last us super long, but should at
least stop the bleeding for now (I'm using # of rows as a proxy for CPU
used, no idea if that's the best proxy but it's at least easy to
measure). We could follow up afterwards to do additional filtering based
on what caseload categories are enabled in a state's outliers_config,
which may or may not also be helpful.

@lilidworkin @ageiduschek @ethan-oro FYI

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 71314740f00724f8c6c6f79522b59eb7700fbefb