[Testing] Cache inspection of pipeline transform signatures (Recidiviz/recidiviz-data#38783)
## Description of the change

**TLDR: Another 2 minutes faster :fire:**
Running: `pytest -k pipeline_integration
recidiviz/tests/ingest/direct/regions`
```
Last PR: 332.25s (0:05:32)
This PR: 198.43s (0:03:18)
```

When profiling the pipelines about _half_ of our pipeline runtime was
being used for `get_function_arguments`. This function calls `inspect`
on beam transforms in order to correctly serialize everything for
workers. It is my understanding this is exacerbated by lambda functions
(as noted in the docstrings)

**For testing:** We cache this function as it should be the same. (Not
sure if this is because of the DirectRunner... you would have to think
Dataflow would cache this, right?)

**For prod:** I made a TODO to get rid of lambda function transforms.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| **Type: Feature** | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 135c07166b68a36b62d0011db833571c14d84686