[looker] Add state entity custom fields (Recidiviz/recidiviz-data#38252)
## Description of the change

in generating the state entity lookml views based solely on a table's
schema, we are deleting many custom dimensions and measures that have
been added lookml view files. to combat that, define
`StateEntityLookMLFieldProvider` that can be used by the
entity_field_builders to look up custom fields for any table.

for custom fields we still want to validate that the underlying field
it's referencing is part of the table schema. it can be tricky to parse
out which schema fields are being referenced by a lookml field once the
field has been created so i took the approach to validate the schema
fields before the lookml field creation.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#23292

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 239fd18195c3724486d1de2ac1120d9b06cb49e2