[Justice Counts] [Database Migration] Surface Number of New Datapoints from Spreadsheet Upload to Uploaded Files Page (Recidiviz/recidiviz-data#38737)
## Description of the change

This PR addresses a CSG request to display the number of new datapoints
on the uploaded files page. Note that this change only affects the
backend; the frontend will later add a new column for this data in the
Uploaded Files page.

Here, a "new datapoint" is defined as any datapoint that has not been
updated with a new value. This information is required for stakeholder
reporting.

The PR implements the following:

- Adds a new column, num_new_datapoints, to the spreadsheet table.
- Calculates the number of new datapoints during the bulk upload process
and stores the result in the new column.
- Surfaces the data in the API response for uploaded files, using the
num_new_datapoints key.

I've tested this change locally as well as updated existing unit tests.

## Related issues

Closes Recidiviz/recidiviz-data#38695

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 2d5ce967bd3d958c9be9ebad3aa3bd81a47ef410