[Tasks][US_TX] Export tasks to dashboard (Recidiviz/recidiviz-data#38837)
## Description of the change

This PR introduces a view to collect the assessment and contact tasks in
TX to be exported to the dashboard.

This PR was tested by running the query manually, exporting the data to
GCS, and running the ETL delegate locally via
```
 py -m recidiviz.tools.workflows.run_etl_from_local_branch  --filename "us_tx_supervision_tasks_record.json" --state_code "US_TX"
```

It was also tested in combination with
https://github.com/Recidiviz/recidiviz-dashboards/pull/7478 during
development. The new typed appeared in the dash.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#37638

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 48b3fd10f672ef5e3673bf0f0340df85206fb73c