[Validation] Add validation surfacing users with upcoming blocks (Recidiviz/recidiviz-data#38435)
## Description of the change

Adds a validation that surfaces any users who have a block date in the
future. These are users who existed in the product roster but were not
included in the latest roster sync file. This validation allows review
of these cases so that the block can be removed if needed before it goes
into effect. For example, in some cases, a user should not actually be
blocked (e.g., they were mistakenly left out of the roster sync file or
they were promoted to leadership).

I figured we only need to select from `user_override` since the `roster`
record will have been deleted already.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#36604

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f71642e1af27f22a6df22e0741d1cd53ce768bbd