[US_AZ] Correct metadata tab name and description (Recidiviz/recidiviz-data#38866)
## Description of the change

Makes the metadata tab name and description match the criteria name for
mandatory literacy criteria. This makes it so that everyone who is
almost eligible for DTP will appear in the tool.

We can see that this will be the case because after this change, there
are no rows in the opportunity record where the metadata tab name is
`NULL`:

```
SELECT * FROM `recidiviz-staging.elise_workflows_views.us_az_approaching_acis_or_recidiviz_dtp_request_record_materialized` WHERE is_almost_eligible IS TRUE AND metadata_tab_name IS NULL LIMIT 1000
```

I don't know how to test that this will flow through as expecting to the
frontend... @ryanrecidiviz can you test off of this branch, or guide me
on how to make sure this fixes the problem?

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: ac64499cbc804a47508925d9edb24a5b6cd2e0c1