[US_TX] Adding small fix to contacts TES criteria  (Recidiviz/recidiviz-data#38873)
## Description of the change

This PR changes two things. Essentially, there was a bug where a contact
was being counted too early in the contact periods. (Imagine we have one
contact on Jan 10, and the person started supervision on Jan 1. Under
the current query there would be two periods Jan 1 - Jan 10 and Jan
10-Enddate. Only the second period should count the contact, not the
first.):
1. Corrects the last contact done date (there was a bug where a contact
was being counted for even though it was at the end of a contact period)
2. Corrects the meets criteria flag to be false for the period as well
## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 6e4ddea6fd397583d7c7b1b5b847fd6f4d62f6d0