[US_AZ] Fix in Risk Release Assessment criteria (Recidiviz/recidiviz-data#38750)
## Description of the change

[US_AZ] Fix in Risk Release Assessment criteria. We will now see periods
where people are ineligible and the reason (assessment_level) they were
ineligible. This will be a lot easier to interpret for anyone.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 36098a8366717368134dfe2eec9a964da5c26742