[kpis] add `column_hydration_score` to `normalized_state_hydration_live_snapshot` (Recidiviz/recidiviz-data#38722)
## Description of the change

this PR adds `column_hydration_score` column to the
`normalized_state_hydration_live_snapshot` table. i had originally
thought it would be easier to do this using `column_hydration_counts` so
we could recalculate this overtime; however, after considering how to
handle schema changes (or changes to what columns are "optional"), i
opted to have this computed at the time when we are sure that the schema
of the `normalized_state` will always match the our query.

i still retained `column_hydration_counts` (repeated record of
`column_name -> non-null counts`) to have as an archive since it's kind
of nice, or in case it turns out we figure out how to better account for
historical changes

plan is to drop existing `normalized_state_hydration_live_snapshot`
before the deploy

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of https://github.com/Recidiviz/recidiviz-data/issues/37632

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 8df67315a60da63435c50dc6d48483dbf7bcc2dd