[US_UT] Temporarily remove half time criteria (Recidiviz/recidiviz-data#38966)
## Description of the change

[US_UT] Temporarily remove half time criteria. This criteria is not yet
hydrated since data is not flowing through sessions yet, so we will
remove this now so we could have a bug bash with actual data on the
tool. We will also have to revert this PR before launch next week.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 44d4055f04090928ff11e3a4f8d2e4f15ffec0f3