[Justice Counts] Add State Codes for U.S Territories (Recidiviz/recidiviz-data#38912)
## Description of the change

Adds State Codes for U.S Territories so that we can create Justice
Counts agencies in those area. This change was requested by CSG.

I've tested this change end to end with this
[frontend](https://github.com/Recidiviz/justice-counts/pull/1684) PR by
creating a new agency in each of the territories.

## Related issues

Closes Recidiviz/recidiviz-data#38911

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 91113151663b1c78f5d9fb2e30429ec2673c5a2f