[CI] Upgrade runners to ubuntu-latest (Recidiviz/recidiviz-data#38985)
## Description of the change

`ubuntu-20.04` is being deprecated at the end of the month. Github has
scheduled the following brownout periods to notify users of the upcoming
deprecation:

* March 4 14:00 UTC – 22:00 UTC
* March 11 13:00 UTC – 21:00 UTC
* March 18 13:00 UTC – 21:00 UTC
* March 25 13:00 UTC – 21:00 UTC

Branches will need to be rebased on top of this change in order for
workflow runs to succeed.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f8190f8f3c50d4332ec6c108d46bfadd250447e5